+++ tests/integration/test_workflow.py
@@ -6,7 +6,7 @@ import pyunicore.client as uc_client
 import pyunicore.credentials as uc_credentials
 
 
-class TestBasic(unittest.TestCase):
+class TestWorkflow(unittest.TestCase):
     def setUp(self):
         pass