+++ CHANGES.md
@@ -16,6 +16,17 @@ Version 2.0.0 (MMM DD, 20yy)
  - use httpx instead of requests as underlying HTTP library for both
    sync and async APIs
 
+Version 1.3.7 (Mar 17, 2026)
+----------------------------
+ - add helper code for uftp authserver
+ - fix build issues with new setuptools
+
+Version 1.3.6  (Dec 18, 2025)
+-----------------------------
+ - CLI: add bearer-token authentication method
+ - CLI: add sshkey authentication method
+ - CLI: resolve ${...} environment values in properties from config file
+ - fix: using non-EdDSA keys did not correctly set the JWT signing algorithm
 
 Version 1.3.5 (Sep 8, 2025)
 ---------------------------